Abstract canvas modification into Circle class
